pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-ai"
version = "0.1.0"
description = "Natural language to shell command translator"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
cli-ai = "cli_ai.main:main"

[tool.setuptools.packages.find]
include = ["cli_ai*"]
